docs: Update all file header documentation
