+++ compiler/hash-alloc/src/brick.rs
@@ -1,6 +1,6 @@
 //! Contains a [`Box`]-like implementation for allocating values within a [`Wall`].
 //!
-//! All rights reserved 2021 (c) The Hash Language authors
+//! All rights reserved 2022 (c) The Hash Language authors
 
 use crate::Wall;
 use core::fmt;
