+++ compiler/hash-alloc/src/collections/mod.rs
@@ -1,7 +1,7 @@
 //! Various collection types that store data inside the memory arena provided by
 //! [`Wall`](super::Wall).
 //!
-//! All rights reserved 2021 (c) The Hash Language authors
+//! All rights reserved 2022 (c) The Hash Language authors
 
 pub mod row;
 pub mod string;
