+++ compiler/hash-ast/src/ident.rs
@@ -1,3 +1,6 @@
+//! Hash AST identifier storage utilities and wrappers.
+//!
+//! All rights reserved 2022 (c) The Hash Language authors
 use dashmap::DashMap;
 
 use fnv::FnvBuildHasher;
