+++ compiler/hash-ast/src/keyword.rs
@@ -1,6 +1,6 @@
 //! Hash Compiler language keyword definitions
 //!
-//! All rights reserved 2021 (c) The Hash Language authors
+//! All rights reserved 2022 (c) The Hash Language authors
 
 use std::fmt;
 use strum_macros::AsRefStr;
