--- compiler/hash-ast/src/storage.rs
@@ -1,6 +1,0 @@
-use hash_alloc::Castle;
-use lazy_static::lazy_static;
-
-lazy_static! {
-    pub static ref STATIC_CASTLE: Castle = Castle::new();
-}
