+++ compiler/hash-ast/src/tree.rs
@@ -1,3 +1,6 @@
+//! AST visualisation utilities.
+//!
+//! All rights reserved 2022 (c) The Hash Language authors
 use std::convert::Infallible;
 use std::iter;
