+++ compiler/hash-ast/src/visitor.rs
@@ -1,6 +1,6 @@
 //! Visitor implementation for [crate::ast] nodes.
 //!
-//! All rights reserved 2021 (c) The Hash Language authors
+//! All rights reserved 2022 (c) The Hash Language authors
 use crate::ast;
 use std::convert::Infallible;
