+++ compiler/hash-interactive/src/command.rs
@@ -1,6 +1,6 @@
 //! Hash interactive mode commands
 //
-// All rights reserved 2021 (c) The Hash Language authors
+// All rights reserved 2022 (c) The Hash Language authors
 
 use hash_reporting::errors::InteractiveCommandError;
