+++ compiler/hash-interactive/src/lib.rs
@@ -1,6 +1,6 @@
 //! Main module for Hash interactive mode.
 //
-// All rights reserved 2021 (c) The Hash Language authors
+// All rights reserved 2022 (c) The Hash Language authors
 
 mod command;
