+++ compiler/hash-pipeline/src/fs.rs
@@ -1,6 +1,6 @@
 //! Hash Compiler filesystem utility functions
 //!
-//! All rights reserved 2021 (c) The Hash Language authors
+//! All rights reserved 2022 (c) The Hash Language authors
 use hash_reporting::reporting::{
     Report, ReportBuilder, ReportCodeBlock, ReportElement, ReportKind, ReportNote,
 };
