+++ compiler/hash-source/src/lib.rs
@@ -1,5 +1,6 @@
-pub mod location;
-
+//! Hash compiler source location definitions.
+//!
+//! All rights reserved 2022 (c) The Hash Language authors
 use slotmap::new_key_type;
 use std::path::Path;
 
@@ -21,3 +22,5 @@ pub trait SourceMap {
     fn path_by_id(&self, source_id: SourceId) -> &Path;
     fn contents_by_id(&self, source_id: SourceId) -> &str;
 }
+
+pub mod location;
