+++ compiler/hash-typecheck/src/traits.rs
@@ -1,3 +1,4 @@
+//! All rights reserved 2022 (c) The Hash Language authors
 use crate::{
     error::{Symbol, TypecheckError, TypecheckResult},
     storage::{GlobalStorage, SourceStorage},
