+++ compiler/hash-utils/src/testing.rs
@@ -1,3 +1,4 @@
+//! All rights reserved 2022 (c) The Hash Language authors
 use std::path::PathBuf;
 
 #[derive(Debug, Clone)]
