+++ compiler/hash/src/main.rs
@@ -1,6 +1,6 @@
 //! Main module.
 //
-// All rights reserved 2021 (c) The Hash Language authors
+// All rights reserved 2022 (c) The Hash Language authors
 
 #![feature(panic_info_message)]
